Change from pre/post snapshot to just a list of snapshots (#47)
Instead of using pre/post snapshot pairs, a DFU package now just consists of a series of snapshots. There must be at least 2 snapshots, which are then used as the pre/post points.

The CLI has been updated to remove `dfu begin` and `dfu end` to replace with just `dfu snap`.

Right now, the pre_snapshot id is always hardcoded to 0, and -1 respectively. A future PR will make this configurable