Print out stderr/stdout when a subprocess fails (#86)
When a git command fails, the subprocess exception is currently unhandled. Not only does that create an ugly stack trace in the terminal, but importantly it doesn't help explain why. This is because the output is captured.

This PR creates a handle_errors decorator, which will display the subprocess error, and then exit(1).

Additionally, the CI scripts are updated to the latest dependencies