Store file permissions in acl.txt (#90)
Git doesn't track file ownership. It only detects whether something is executable. In order to properly handle this, we need to do this ourselves.

This PR introduces an acl.txt file, which contains the permissions to apply to the file when saving to the file system. Getting this to work properly ended up changing many things about the codebase, with specific care needed to things like symlinks that were otherwise not handled well.

The new algorithm is this:
1. Copy the files to the playground
 2. chown and chmod --recursive to make sure the current git user can see them
 3. Save the original file permissions from the snapshot in acl.txt
 4. When applying a patch, copy the files over, then use chmod/chown to set them to the correct value from acl.txt